+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- LoadingText updates text from Text and/or TextMeshPro
+
 ## [3.0.0] - 2024-10-03
 ### Added
 - LoadScene function